Refactor bubblesort implementation
Previous implementation was done in a very procedural way. Implementing it as a function would make it easy to be used as a personal algorithm library. If we want to copy the bubble sort implementation, we just need to copy line 6 to line 21 to our c++ files and call bubble_sort(our_vector);

I believe this is called better practice