stack: full implementation of stack in c
We are going to implement the basic features of stack
using linked list in C.
The basic features of stack includes PUSH,POP and PEEK.
In PUSH operation we are adding and element into the stack.
In POP operation we are deleting an element from the stack.
In PEEK operation we are printing the top most element of
the stack.
We are also using some more functions in our program,
'is_empty' function allows us to check whether the stack
is empty.
we are using a function called "refresh", the basic function
of refresh is to clear the screen.

fixes: #920 (https://github.com/OpenGenus/cosmos/issues/920)

Signed-off-by: SATYAM KUMAR <satyamvats5@gmail.com>