Add Bazel WORKSPACE and a single BUILD file
GitHub issue: OpenGenus/cosmos#1702

The WORKSPACE file is global for Cosmos and lets Bazel know where the project
root is. When Bazel builds the libraries in Cosmos, it creates some directories
that start with "bazel-" in the project root, so I wildcard added these to
.gitignore.

Just to try this out, I added a BUILD file for the red-black tree library that
builds each language's implementation. Each library also has a unit test
associated with it, to demonstrate how we can run `bazel test //...` and
every test in Cosmos will be run.