Multi Layer Perceptron
Implementation of supervised learning algorithm called Multi-Layer Perceptron (MLP) by extending Artificial Neural Networks (ANN) and using keras.
The Keras Python library for deep learning focuses on the creation of models as a sequence of layers. The simplest model is defined in the Sequential class which is a linear stack of Layers.