Update dijkstra.js
`Array(noOfVertices).fill(Array(noOfVertices).fill(0))` does not create an nxn matrix of zeroes. It creates an array of n references to the same array of n zeroes (that is to say, only 2 arrays are ever being created, rather than n+1). I also put a default value on the `Graph#dijkstra` method so it can be invoked without an argument.