Implementation of Recurrent Neural Network (RNN) from scratch
Implementation of Recurrent Neural Network (RNN) from scratch (char-rnn).
All the RNN code is in rnn.py.
train.py is used for defining an instance of the RNN and training it.
generate.py is for generating characters.
uwv.pkl contains weights for the trained model. If you run train.py, the model will load the weights from this file and continue training from there.