Create optimized segment tree implementation
Changes: Adding a file for the code of segment tree which always require only 2N space always whether the size of the array is the power of 2 or not. The code has bottom-up approach for sum queries and updating values so that we can achieve it using iteration and the analysis is much easier.