Correct Grammar and clarify algorithm
I was very unimpressed by the initial explanation for the algorithm and didn't thing it did the sieve justice. So I've gone ahead and written a much clearer and more grammatically correct version which includes additional resources and examples to clarify the algorithm, it's logic, why it works, and much much more.