Create Added _new _css _padding.html
solved issue #6664 


In this example, the CSS padding property is used to add space around the content of a div element. The padding property can take four values: top, right, bottom, and left. In this case, the padding is set to 20 pixels for all four sides for the container div.

The unique-padding div has a different padding value of 10px 20px 30px 40px. This means that the top padding is 10 pixels, the right padding is 20 pixels, the bottom padding is 30 pixels, and the left padding is 40 pixels.